test/topology: Move helpers to get tablet replicas to pylib
These are very useful and will be used across different test files soon
